requirements.txt
langchain==0.0.175
llama-index==0.6.9
openai==0.27.4
python-dotenv==1.0.0
PyPDF2==3.0.1
EbookLib==0.18
docx2txt==0.8
html2text==2020.1.16
rarfile==4.0
py7zr==0.20.5
